Move the heed types into their own module
