+++ meilisearch-core/src/heed_types/mod.rs
@@ -1,0 +1,7 @@
+mod cow_set;
+mod documents_ids;
+mod postings;
+
+pub use self::cow_set::CowSet;
+pub use self::documents_ids::{DocumentsIds, DiscoverIds};
+pub use self::postings::{Postings, PostingsCodec};
