+++ meilisearch-core/src/store/mod.rs
@@ -1,6 +1,4 @@
-mod cow_set;
 mod docs_words;
-mod documents_ids;
 mod documents_fields;
 mod documents_fields_counts;
 mod facets;
@@ -12,11 +10,9 @@ mod synonyms;
 mod updates;
 mod updates_results;
 
-pub use self::cow_set::CowSet;
 pub use self::docs_words::DocsWords;
 pub use self::documents_fields::{DocumentFieldsIter, DocumentsFields};
 pub use self::documents_fields_counts::{DocumentFieldsCountsIter, DocumentsFieldsCounts, DocumentsIdsIter};
-pub use self::documents_ids::{DocumentsIds, DiscoverIds};
 pub use self::facets::Facets;
 pub use self::main::Main;
 pub use self::postings_lists::PostingsLists;
@@ -26,15 +22,10 @@ pub use self::synonyms::Synonyms;
 pub use self::updates::Updates;
 pub use self::updates_results::UpdatesResults;
 
-use std::borrow::Cow;
 use std::collections::HashSet;
-use std::convert::TryInto;
-use std::{mem, ptr};
 
 use heed::Result as ZResult;
-use heed::{BytesEncode, BytesDecode};
 use meilisearch_schema::{IndexedPos, FieldId};
-use sdset::{Set, SetBuf};
 use serde::de::{self, Deserialize};
 use zerocopy::{AsBytes, FromBytes};
 
@@ -43,7 +34,8 @@ use crate::database::{MainT, UpdateT};
 use crate::database::{UpdateEvent, UpdateEventsEmitter};
 use crate::serde::Deserializer;
 use crate::settings::SettingsUpdate;
-use crate::{query_builder::QueryBuilder, update, DocIndex, DocumentId, Error, MResult};
+use crate::{query_builder::QueryBuilder, update, DocumentId, Error, MResult};
+use crate::heed_types::{DocumentsIds, CowSet};
 
 type BEU32 = zerocopy::U32<byteorder::BigEndian>;
 type BEU64 = zerocopy::U64<byteorder::BigEndian>;
@@ -78,87 +70,6 @@ impl DocumentFieldStoredKey {
             docid: BEU32::new(docid.0),
             field_id: BEU16::new(field_id.0),
         }
-    }
-}
-
-#[derive(Default, Debug)]
-pub struct Postings<'a> {
-    pub docids: Cow<'a, Set<DocumentId>>,
-    pub matches: Cow<'a, Set<DocIndex>>,
-}
-
-pub struct PostingsCodec;
-
-impl<'a> BytesEncode<'a> for PostingsCodec {
-    type EItem = Postings<'a>;
-
-    fn bytes_encode(item: &'a Self::EItem) -> Option<Cow<'a, [u8]>> {
-        let u64_size = mem::size_of::<u64>();
-        let docids_size = item.docids.len() * mem::size_of::<DocumentId>();
-        let matches_size = item.matches.len() * mem::size_of::<DocIndex>();
-
-        let mut buffer = Vec::with_capacity(u64_size + docids_size + matches_size);
-
-        let docids_len = item.docids.len() as u64;
-        buffer.extend_from_slice(&docids_len.to_be_bytes());
-        buffer.extend_from_slice(item.docids.as_bytes());
-        buffer.extend_from_slice(item.matches.as_bytes());
-
-        Some(Cow::Owned(buffer))
-    }
-}
-
-fn aligned_to(bytes: &[u8], align: usize) -> bool {
-    (bytes as *const _ as *const () as usize) % align == 0
-}
-
-fn from_bytes_to_set<'a, T: 'a>(bytes: &'a [u8]) -> Option<Cow<'a, Set<T>>>
-where T: Clone + FromBytes
-{
-    match zerocopy::LayoutVerified::<_, [T]>::new_slice(bytes) {
-        Some(layout) => Some(Cow::Borrowed(Set::new_unchecked(layout.into_slice()))),
-        None => {
-            let len = bytes.len();
-            let elem_size = mem::size_of::<T>();
-
-            // ensure that it is the alignment that is wrong
-            // and the length is valid
-            if len % elem_size == 0 && !aligned_to(bytes, mem::align_of::<T>()) {
-                let elems = len / elem_size;
-                let mut vec = Vec::<T>::with_capacity(elems);
-
-                unsafe {
-                    let dst = vec.as_mut_ptr() as *mut u8;
-                    ptr::copy_nonoverlapping(bytes.as_ptr(), dst, len);
-                    vec.set_len(elems);
-                }
-
-                return Some(Cow::Owned(SetBuf::new_unchecked(vec)));
-            }
-
-            None
-        }
-    }
-}
-
-impl<'a> BytesDecode<'a> for PostingsCodec {
-    type DItem = Postings<'a>;
-
-    fn bytes_decode(bytes: &'a [u8]) -> Option<Self::DItem> {
-        let u64_size = mem::size_of::<u64>();
-        let docid_size = mem::size_of::<DocumentId>();
-
-        let (len_bytes, bytes) = bytes.split_at(u64_size);
-        let docids_len = len_bytes.try_into().ok().map(u64::from_be_bytes)? as usize;
-        let docids_size = docids_len * docid_size;
-
-        let docids_bytes = &bytes[..docids_size];
-        let matches_bytes = &bytes[docids_size..];
-
-        let docids = from_bytes_to_set(docids_bytes)?;
-        let matches = from_bytes_to_set(matches_bytes)?;
-
-        Some(Postings { docids, matches })
     }
 }
