+++ meilisearch-core/src/update/documents_addition.rs
@@ -14,7 +14,8 @@ use crate::database::{UpdateEvent, UpdateEventsEmitter};
 use crate::facets;
 use crate::raw_indexer::RawIndexer;
 use crate::serde::Deserializer;
-use crate::store::{self, DocumentsFields, DocumentsFieldsCounts, DiscoverIds};
+use crate::store::{self, DocumentsFields, DocumentsFieldsCounts};
+use crate::heed_types::DiscoverIds;
 use crate::update::helpers::{index_value, value_to_number, extract_document_id};
 use crate::update::{apply_documents_deletion, compute_short_prefixes, next_update_id, Update};
 use crate::{Error, MResult, RankedMap};
